pyproject.toml
[project]
name = "dcepersona"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "openai>=2.6.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "ruff>=0.14.8",
    "typer>=0.15.0",
]

[tool.ruff]
line-length = 128
target-version = "py312"

# 自動修正（ruff check --fix）で有効化されるルール
fix = true

# 無視したいルール例（必要なければそのままでOK）
lint.ignore = [
    "E501",  # line length（Black を使うなら無視する人が多い）
]

# 使用するルールセット（必要に応じて追加）
lint.select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "W",  # Warning
    "B",  # Bugbear
    "C90", # mccabe complexity
]

[tool.ruff.format]
# Black 互換の formatter を使う
quote-style = "double"
indent-style = "space"
line-ending = "lf"
